requirements.txt
dj-database-url
Django==1.3.1
djangohelpers==0.8
psycopg2
MySQL-python
pysqlite
amqplib==1.0.2
django-debug-toolbar==0.9.4
django-htauth
django-actionkit-client
django-kombu==0.9.4
django-picklefield==0.1.9
ordereddict==1.1
requests==0.14.0
django-celery==2.4.2
celery==2.4.6
